ai-chip-mock-servers/mock-server/requirements.txt
flask
flask-cors
influxdb==5.3.1
influxdb-client==1.39.0
apscheduler
flask-apscheduler
python-dotenv